requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
psycopg[binary]>=3.1
python-dotenv>=1.0
email-validator>=2.1
../jarvis_contracts
aiohttp>=3.9.0
